refactor: rename module name
napi is no longer frontend only
